rename reg module
